==> 0 CodingID.ipynb <==
import pandas as pd

path = '/Users/fan/Desktop/Fullhouse/WCM/HospitalManagement/OrderAnalysis-master/'

# Order

#order_task_all = pd.read_csv(path + 'order_2012_2018_task_pull2_all.csv', header=0)
#order = pd.read_csv(path + 'order_taskocc_2012_2018.csv', header=0)
order = pd.read_csv(path + 'order_EDDC_2012_2018_cv3_notcancelled.csv', header=0)
order.head()

#order['TaskStatusCode'].value_counts()

order['enterrole'].value_counts()

order.info()

order['clientvisitguid'] = order['clientvisitguid'] - 9000000000000000
order['clientguid'] = order['clientguid'] - 9000000000000000
order['userguid'] = 10000000000000000 - order['userguid']

print(order['clientvisitguid'].min())
print(order['clientvisitguid'].max())

order[order['clientvisitguid']== order['userguid']]

order.to_csv(path + 'order_EDDC_2012_2018_cv3_notcancelled_new.csv', index=False)

# client location

location = pd.read_csv(path + 'clientlocation_2012_2018_pull2.csv', header = 0)
location.head()

location.info()

location['clientvisitguid'] = location['clientvisitguid'] - 9000000000000000

print(location['clientvisitguid'].min())
print(location['clientvisitguid'].max())

location.to_csv(path + 'clientlocation_2012_2018_pull2_new.csv', index=False)

# client visit

clientvisit = pd.read_csv(path + 'clientvisit_2012_2018_pull2.csv', header = 0)
clientvisit.head()

clientvisit.info()

clientvisit['ClientVisitGUID'] = clientvisit['ClientVisitGUID'] - 9000000000000000

print(clientvisit['ClientVisitGUID'].min())
print(clientvisit['ClientVisitGUID'].max())

clientvisit['ClientGUID'] = clientvisit['ClientGUID'] - 9000000000000000

print(clientvisit['ClientGUID'].min())
print(clientvisit['ClientGUID'].max())

clientvisit[clientvisit['ClientGUID']==clientvisit['ClientVisitGUID']]

clientvisit.to_csv(path + 'clientvisit_2012_2018_pull2_new.csv', index=False)

# demographics

demo = pd.read_csv(path + 'demographic_2012_2018_pull2.csv', header = 0)
demo.head()

demo.info()

demo['clientguid'] = demo['clientguid'] - 9000000000000000

demo['clientvisitguid'] = demo['clientvisitguid'] - 9000000000000000

demo.to_csv(path + 'demographic_2012_2018_pull2_new.csv', index=False)

# diagnosis

diag = pd.read_csv(path + 'diag_2012_2018_pull2.csv')
diag.head()

diag.info()

diag['ClientGUID'] = diag['ClientGUID'] - 9000000000000000

diag['ClientVisitGUID'] = diag['ClientVisitGUID'] - 9000000000000000

diag['userguid'] = 10000000000000000 - diag['userguid']

diag[diag['ClientGUID']== diag['ClientVisitGUID']]

diag[diag['userguid']== diag['ClientVisitGUID']]

diag[diag['ClientGUID']== diag['userguid']]

diag.to_csv(path + 'diag_2012_2018_pull2_new.csv', index=False)

# EDDC_ordertask_2012_2018_cluster_MED.xlsx

ordertask_med = pd.read_excel(path + 'EDDC_ordertask_2012_2018_cluster_MED.xlsx')
ordertask_med.head()

ordertask_med.info()

ordertask_med['clientvisitguid'] = ordertask_med['clientvisitguid'] - 9000000000000000
ordertask_med['clientvisitguid']

ordertask_med.to_excel(path + 'EDDC_ordertask_2012_2018_cluster_MED_new.xlsx', index=False)

# EDDC_order_2012_2018_cluster_MED.xlsx

order_med = pd.read_excel(path + 'EDDC_order_2012_2018_cluster_MED.xlsx', sheet_name = 'EDDC_order_2012_2018_cluster')
order_med.head()

order_med.info()

order_med['clientvisitguid'] = order_med['clientvisitguid'] - 9000000000000000
order_med['clientvisitguid']

order_med.to_excel(path + 'EDDC_order_2012_2018_cluster_MED_new.xlsx', index=False)

# Insurance

insurance = pd.read_csv(path + 'insurance_2012_2018_pull2.csv', header = 0)
insurance.head()

insurance.info()

insurance.columns = ['clientvisitguid', 'Financial Group']

insurance['clientvisitguid'] = insurance['clientvisitguid'] - 9000000000000000

insurance.to_csv(path + 'insurance_2012_2018_pull2_new2.csv', index=False)
