==> 3.1 ComorbidityScore_JSONtoCSV.ipynb <==
import json
import csv
import pandas as pd

## extract diagnosis from appt

fh = open('/Users/fan/Desktop/Fullhouse/WCM/HospitalManagement/OrderAnalysis-master/2_EDDC_HF_2012_2018_ouput.json','r')

data = json.load(fh)

len(data)

for pid in data:
    if pid == '1285016600270':
        for date in data[pid]['appt']: 
            print(pid, data[pid]['appt'][date]['diag'])

with open('/Users/fan/Desktop/Fullhouse/WCM/HospitalManagement/OrderAnalysis-master/2_diag_original.csv', 'w+') as f:
    for pid in data:
        for date in data[pid]['appt']: 
            print(pid, data[pid]['appt'][date]['diag'], file = f, sep = ' ')

diag = pd.read_csv('/Users/fan/Desktop/Fullhouse/WCM/HospitalManagement/OrderAnalysis-master/2_diag_original.csv', sep = '  ', names = ['ID'])

diag

diag_1 = diag.ID.str.split(expand = True)

#diag.columns = ['oldname']
diag_1.columns = ['ID', 'icd1', 'icd2', 'icd3', 'icd4', 'icd5', 'icd6', 'icd7', 'icd8', 'icd9', 'icd10', 
                  'icd11', 'icd12', 'icd13', 'icd14', 'icd15', 'icd16', 'icd17', 'icd18', 'icd19', 'icd20', 
                 'icd21', 'icd22', 'icd23', 'icd24', 'icd25', 'icd26', 'icd27']

diag_1

for i in range(len(diag_1.columns)):
    diag_1.iloc[:,i] = diag_1.iloc[:,i].str.strip('[')
    diag_1.iloc[:,i] = diag_1.iloc[:,i].str.strip(']')
    diag_1.iloc[:,i] = diag_1.iloc[:,i].str.strip(',')
    diag_1.iloc[:,i] = diag_1.iloc[:,i].str.strip("''")

diag_1

diag_1['ID'] = diag_1['ID'].astype(int)

diag_2 = pd.melt(frame=diag_1, id_vars='ID')

diag_2 = diag_2.drop(['variable'], axis='columns')

diag_2

nan_value = float("NaN")
diag_2.replace("nan", nan_value, inplace = True)
diag_2.replace("", nan_value, inplace = True)
diag_2.dropna(subset = ["value"], inplace = True)

diag_2

diag_2['value'].nunique()

diag_2[diag_2['ID'] == 806335300270]

diag_3 = diag_2.drop_duplicates(subset=('ID', 'value'))

diag_3

diag_3[diag_3['ID'] == 806335300270]

diag_2.groupby(['ID'])['value'].unique()

diag_2.to_csv('/Users/fan/Desktop/Fullhouse/WCM/HospitalManagement/Data/2_diag_withduplicates.csv', index=None)

diag_3.to_csv('/Users/fan/Desktop/Fullhouse/WCM/HospitalManagement/Data/2_diag_withoutduplicates.csv', index=None)
