==> 6.1 ProcessSeq_beforeClustering.ipynb <==
import pandas as pd

path='/Users/fan/Desktop/Fullhouse/WCM/HospitalManagement/OrderAnalysis-master/'

## using the seq.xlsx which got the printout from getV_fromCSV.py (line 332)

seq = pd.read_excel(path + 'seq.xlsx', sep = ' ', header=None)

seq.head()
seq.columns= ['ID']
seq.head()

seq = seq.ID.str.split(expand = True)

seq.info()

len(seq.columns)

for i in range(len(seq.columns)):
    seq.iloc[:,i] = seq.iloc[:,i].str.strip('[')
    seq.iloc[:,i] = seq.iloc[:,i].str.strip(']')
    seq.iloc[:,i] = seq.iloc[:,i].str.strip(',')
    seq.iloc[:,i] = seq.iloc[:,i].str.strip("''")

len(seq.index)

seq.head()

seq.to_csv(path + "seq_clustering.csv", index=None)
