==> 6.2 TrainTestSplit.ipynb <==
Within the dataset Data2(EDDC_HF_data_filtered_OS.json) (1651 clientvisitguids), the clientvisitids appearing in the Yiye's file(EDDC_order_2012_2018_cluster_MED_task_YY.xlsx) (1412 clientvisitguids) will be treated as training data, and the rest of clientvisitids will be treated as testing data

import json
import csv
import pandas as pd

path = '/Users/fan/Desktop/Fullhouse/WCM/HospitalManagement/OrderAnalysis-master/'

order_task = pd.read_csv(path + 'order_2012_2018_task_all_new.csv', header=0)
#location = pd.read_csv(path + 'Data/clientlocation_2012_2018_pull2_new.csv', header = 0)
#clientvisit = pd.read_csv(path + 'Data/clientvisit_2012_2018_pull2_new.csv', header = 0)
#demo = pd.read_csv(path + 'Data/demographic_2012_2018_pull2_new.csv', header = 0)
#diag = pd.read_csv(path + 'Data/diag_2012_2018_pull2_new.csv')

## all ids
alldata = pd.read_excel(path + 'EDDC_order_2012_2018_cluster_MED_task_yifanfinal_1651.xlsx', header=0)
alldata.head()

train_data = alldata[alldata.iloc[:,0].isin(train_id)]
train_data

train_data.to_excel(path + 'train_data.xlsx', index=False)

test_data = alldata[alldata.iloc[:,0].isin(test_id)]
test_data

test_data.to_excel(path + 'test_data.xlsx', index=False)

fh = open(path + 'Data/EDDC_HF_data_filtered_OS.json','r')

data = json.load(fh)

with open(path + 'Data/allpatients.csv', 'w+') as f:
    for pid in data:
            print(pid, file = f, sep = ' ')

Load all the patient ids

allpatients = list(alldata['clientvisitguid'])

allpatients

allpatients = pd.read_csv(path + 'Data/allpatients.csv', header=None, names=['ID'])

allpatients

allpatients['ID'].nunique()

Load train patient ids

## File Yiye created before for order_task
ordertask_med = pd.read_excel(path + 'EDDC_order_2012_2018_cluster_MED_task_YY.xlsx', sheet_name = 'EDDC_order_2012_2018_cluster', header = 0)

ordertask_med.head()

ordertask_med[~ordertask_med['clientvisitguid'].isin(allpatients)]['clientvisitguid'].nunique()

## test data number: 295
test_id = alldata[~alldata['clientvisitguid'].isin(ordertask_med['clientvisitguid'])]['clientvisitguid'].unique()

## train data number: 1409
train_id = ordertask_med[ordertask_med['clientvisitguid'].isin(allpatients)]['clientvisitguid'].unique()

print(len(test_id))
print(len(train_id))

## find the missing ids in the File Yiye created before for order_task, and delete them from train data
ordertask_med[~ordertask_med['clientvisitguid'].isin(allpatients['ID'])]['clientvisitguid'].unique()

missid_ordertask = list(ordertask_med[~ordertask_med['clientvisitguid'].isin(allpatients['ID'])]['clientvisitguid'].unique())

missid_ordertask

## Export train sequence and test sequence seperately

seq = pd.read_csv(path + "seq_clustering.csv", header = 0)
seq.head()

len(seq)

## the number of seq(1629) is 20 less than that of data2 (1649)
seqid = seq.iloc[:,0].unique()
allpatients[~allpatients['ID'].isin(seqid)].nunique()

train_seq = seq[seq.iloc[:,0].isin(train_id)]
train_seq

train_seq.to_csv(path + "train_seq_clustering.csv", index=None)

test_seq = seq[seq.iloc[:,0].isin(test_id)]
test_seq

test_seq.to_csv(path + "test_seq_clustering.csv", index=None)

#对于每一个clientvisitguid你能看一下有多少order是order set 多少不是 order set吗

order_task

clientvisitguid_order = order_task[order_task['clientvisitguid'].isin(alldata['clientvisitguid'])].groupby(['clientvisitguid'])[['name']].count()

clientvisitguid_order

pd.DataFrame(order_task[order_task['clientvisitguid'].isin(alldata['clientvisitguid'])].
                                        groupby(['clientvisitguid'])[['ordersetname']].apply(lambda x: x.notnull().sum()))

clientvisitguid_orderset

clientvisitguid_order.to_excel(path + '2_clientvisitguid_order_1651.xlsx')
clientvisitguid_orderset.to_excel(path + '2_clientvisitguid_orderset_1651.xlsx')

order_task.info()
