==> 7.1 GetDescriptiveTable_OpioidOrdercount.ipynb <==
import pandas as pd

path='/Users/fan/Desktop/Fullhouse/WCM/'

data=pd.read_csv(path + 'HospitalManagement/OrderAnalysis-master/order_taskocc_2012_2018_new.csv', sep=',')

len(data)

data1 = data[data['TaskStatusCode']!='Canceled']

len(data1)

data2=data1[data1['enterrole'].isin(['PA', 'Resident', 'Nurse', 'NP', 'Physician', 'VerifyChemoTherapy',
       'Fellow', 'Medical Resident', 'Attending Physician',
       'Registered Nurse', 'Nurse Anesthetist'])]

len(data2)

medication = data2[data2['typecode']=='Medication']

len(medication)

medication['clientvisitguid'].nunique()

#medication = medication.drop(['TaskStatusCode', 'clientguid', 'PerformedFromDtm', 'enterrole', 'typecode', 'userguid'], axis='columns')

medication

opioid = pd.read_excel(path + 'Opioid/Opioid Order GUID.xlsx', header=0)

opioid

opioid_d = opioid.loc[:, ['Generic', 'OrderName']]
opioid_d

opioid_d_dropduplicates = opioid_d.drop_duplicates()

opioid_d_dropduplicates['OrderName'].unique()

opioid_d_dropduplicates['OrderName'].nunique()

opioid_guid = opioid.loc[:, ['GUID', 'ImportSource', 'OrderName']]
opioid_guid

medication_opioid_dropduplicates = pd.merge(left=medication, right=opioid_d_dropduplicates, 
                       left_on = 'name', right_on = 'OrderName', how = 'left')

medication_opioid_dropduplicates

medication_opioid_dropduplicates[medication_opioid_dropduplicates['clientvisitguid']==1002827500270]['name'].value_counts()

medication[medication['clientvisitguid']==503942200270]['name'].value_counts()

medication['TaskStatusCode'].value_counts()

medication_opioid_dropduplicates['TaskStatusCode'].value_counts()

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'Fentanyl Intranasal Soln 50 mcg/ml', 'Generic'] ='Fentanyl'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'Fentanyl Intranasal Soln 50 mcg/ml', 'OrderName'] ='fentaNYL Intranasal Soln 50 mcg/ml'

fentaNYL Patch 100 microgram/hr    
fentaNYL Patch  12 microgram/hr

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 12 microgram/hr', 'Generic'] ='Fentanyl'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 12 microgram/hr', 'OrderName'] ='fentaNYL Patch'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 25 microgram/hr', 'Generic'] ='Fentanyl'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 25 microgram/hr', 'OrderName'] ='fentaNYL Patch'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 50 microgram/hr', 'Generic'] ='Fentanyl'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 50 microgram/hr', 'OrderName'] ='fentaNYL Patch'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 75 microgram/hr', 'Generic'] ='Fentanyl'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 75 microgram/hr', 'OrderName'] ='fentaNYL Patch'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 100 microgram/hr', 'Generic'] ='Fentanyl'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'fentaNYL Patch 100 microgram/hr', 'OrderName'] ='fentaNYL Patch'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'Fentanyl Citrate Inj +R+', 'Generic'] ='Fentanyl'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'Fentanyl Citrate Inj +R+', 'OrderName'] ='fentaNYL Citrate Inj +R+'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'PCEA Fentanyl 2 mcg/mL, bupivacaine 0.1 % in 0.9% sodium chloride', 'Generic'] ='Fentanyl'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'PCEA Fentanyl 2 mcg/mL, bupivacaine 0.1 % in 0.9% sodium chloride', 'OrderName'] ='PCEA fentaNYL 2 mcg/mL, bupivacaine 0.1 % in 0.9% sodium chloride'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'Hydrocodone/Acetaminophen 5mg/334mg Liq (per 10mL)', 'Generic'] ='Hydrocodone'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'Hydrocodone/Acetaminophen 5mg/334mg Liq (per 10mL)', 'OrderName'] ='HYDROcodone/Acetaminophen 5mg/334mg Liq (per 10mL)'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'oxyCODONE/Acetaminophen 5/325 mg Tab (Percocet)', 'Generic'] ='Oxycodone'

medication_opioid_dropduplicates.loc[medication_opioid_dropduplicates['name'] == 'oxyCODONE/Acetaminophen 5/325 mg Tab (Percocet)', 'OrderName'] ='Oxycodone/Acetaminophen 5/325 mg Tab (Percocet)'

## check the failure in matching is the capital issue
a = pd.DataFrame(medication_opioid_dropduplicates[medication_opioid_dropduplicates['OrderName'].isnull()==True]['name'].unique())

a.to_excel(path + 'Opioid/a_dropduplicates.xlsx', index=False)

#check
medication_opioid_dropduplicates[medication_opioid_dropduplicates['clientvisitguid']==1184048800270]['name'].value_counts()

#check
medication_opioid_dropduplicates[medication_opioid_dropduplicates['clientvisitguid']==1002827500270].groupby(['Generic', 'TaskStatusCode'])['name'].count()

medication_opioid_dropduplicates.columns

len(medication_opioid_dropduplicates)

medication_opioid_dropduplicates.to_csv(path + 'Opioid/medication_opioid_12feature_dropduplicates_filtered_0305.csv', index=None)

Match order data with opioid drugs

order_opioid_guid_match_filtered = pd.merge(left=medication_opioid_dropduplicates, right=opioid_guid, 
                       on = 'OrderName', how = 'left')

len(order_opioid_guid_match_filtered)

order_opioid_guid_match_filtered.to_csv(path + 'Opioid/medication_opioid_guid_order_match_filtered.csv')

#check
medication_opioid_dropduplicates[(medication_opioid_dropduplicates['clientvisitguid']==1176574600270) & 
                                 (medication_opioid_dropduplicates['Generic']=='Oxycodone') & 
                                 (medication_opioid_dropduplicates['PerformedFromDtm'] == '2018-05-26 17:05:51.000')]

order_opioid_guid_match_filtered[(order_opioid_guid_match_filtered['clientvisitguid']==1176574600270) & 
                                 (order_opioid_guid_match_filtered['Generic']=='Oxycodone') & 
                                 (order_opioid_guid_match_filtered['PerformedFromDtm']=='2018-05-26 17:05:51.000')]

medication_opioid_dropduplicates.groupby(['clientvisitguid', 'Generic', 'TaskStatusCode'])['name'].count()

ordercountby_id_generic_taskstatus=pd.DataFrame(medication_opioid_dropduplicates.groupby(['clientvisitguid', 'Generic', 'TaskStatusCode'])['name'].count())

medication_opioid_dropduplicates['TaskStatusCode'].value_counts()

## check why the 'Rescheduled' records disappear
medication_opioid_dropduplicates[medication_opioid_dropduplicates['TaskStatusCode']=='Rescheduled']

ordercountby_id_generic_taskstatus

ordercountby_id_generic_taskstatus = ordercountby_id_generic_taskstatus.reset_index()

ordercountby_id_generic_taskstatus['TaskStatusCode'].value_counts()

result = ordercountby_id_generic_taskstatus.pivot_table(index=['clientvisitguid', 'Generic'], columns='TaskStatusCode', values='name')

result

result2 = result.reset_index()

result2

result2.to_excel(path + 'Opioid/ordercountby_id_generic_taskstatus_dropduplicates_filtered_0305.xlsx', index=None)
